Added argument 'collectProcessRelatedMetrics'
In some cases exporter and OVS are not in the same namespace
(exporter doesn't have access to files of OVS processes in '/proc')
To exclude wrong and weird exporter's behaviour
we wish to exclude process related metrics.
New command-line argument is boolean and equals 'true' by default.
In order to exclude process related metrics -
launch exporter with '-collectProcessRelatedMetrics=false'